# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WMS — Warehouse Management System API",
        "description": "API REST do sistema de Gestão de Almoxarifado (Laravel 13 + Sanctum SPA).",
        "contact": {
            "name": "WMS",
            "email": "suporte@wms.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local (Docker)"
        }
    ],
    "paths": {
...
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Autenticar usuário (SPA Sanctum cookie)",
                "operationId": "bae68d1c55329ecb6e41f48285840c8f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@wms.local"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "Admin@123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Credenciais inválidas"
                    },
                    "429": {
                        "description": "Muitas tentativas"
                    }
                }
            }
        },
        "/api/v1/me": {
            "get": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Perfil do usuário autenticado (roles + permissões)",
                "operationId": "13b43df400e475a157538f73617b8c3d",
                "responses": {
                    "200": {
                        "description": "Perfil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "Usuário",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Administrador"
                    },
                    "email": {
                        "type": "string",
                        "example": "admin@wms.local"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "administrador"
                        ]
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "products.view"
                        ]
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "session",
                "scheme": "bearer"
            }
        }
    },
...
}